:sparkles: #4 Join 클래스 분리
